requirements.txt
streamlit==1.32.0
streamlit-shadcn-ui==0.1.2
python-dotenv==1.0.0
feedparser==6.0.11
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
anthropic==0.18.1
groq==0.4.2
resend==0.8.0
supabase==2.3.0
postgrest==0.13.0
tweepy==4.14.0
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytube==15.0.0
schedule==1.2.1
python-dateutil==2.8.2
pytz==2024.1
